Signal handling refactor
This is a removal of legacy code for signal handling. This signal
handling was useful for the legacy runner where all tasks were running
in child processes. But now when we use spawners for creating tasks each
with different implementation, we need to handle the signal handling in
the spawners itself.

Signed-off-by: Jan Richter <jarichte@redhat.com>